requirement.txt
pillow
fastapi
moviepy
firebase-admin
uvicorn
sqlalchemy
pycairo
python-multipart
requests
opencv-python